[Feature][James] separated the Side panel into its own component
